requirements.txt
easydict==1.13
h5py==3.11.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
nltk==3.8.1
numpy==1.26.4
open3d==0.18.0
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.2.0
plyfile==1.0.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tokenizers==0.19.1
torch-cluster==1.6.0+pt112cu113
torch_geometric==2.5.0
torch-scatter==2.1.0+pt112cu113
torch-sparse==0.6.16+pt112cu113
torch-spline-conv==1.2.1+pt112cu113
torch-tb-profiler==0.4.3
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.40.0
transforms3d==0.4.1